refactor: modularize MemRegion
